feat(flight): double-precision FlightState position vector (#371) (#388)
* feat(flight): double-precision FlightState position vector (#371)

FlightState::pos_world[3] is upgraded from float[3] to double[3],
closing the silent precision gap at large world-space distances.
At x = 1e5 m the float ULP (~0.0078 m) exceeds the per-step lateral
gravity displacement (~0.001 m); the position never moved. With
double, sub-millimetre increments accumulate correctly at any scale
including planet-radius distances needed by #354 and #355.

Interface changes (primary development — no compat constraints):
- IGravityField::accelWorld and geodeticAltitude params upgraded to
  const double[]; geodeticAltitude return upgraded to double
- CentralGravityField uses double arithmetic internally for the
  direction computation; float return via explicit cast (gravity
  magnitudes ~9.8 m/s² are fine in float)
- FlightIntegrator.cpp: two static_cast<float> at computeAtmosphere
  call sites (deliberate float boundary)
- WorldBroadcaster: setGroundElevationQuery callback upgraded to
  float(double,double); m_entityX/Z atomics upgraded to
  atomic<double> with static_assert(is_always_lock_free);
  cachedEntityX/Z() return double

Wire format unchanged (MsgEntityEntry::pos already double[3]).
kProtocolVersion stays at 1.

Adds precision regression tests in test_flight_integrator and
test_world_broadcaster that would fail against the old float[3].

* fix(flight): fix MSVC C4244 narrowing warnings on Windows build

Two implicit double→float narrowings caught by MSVC C4244
(-Werror equivalent) that GCC/Clang silently accepted:

- test_force_model.cpp: fallTo() returned float but pos_world[1] is
  now double; change return type and all call-site locals to double
- test_admin_console.cpp: setGroundElevationQuery lambda used
  (float, float) params but the callback is float(double, double);
  MSVC fires C4244 when constructing std::function from a narrowing
  lambda; change params to (double, double)

+++ engine/flight/IGravityField.h
@@ -7,17 +7,16 @@ namespace fl {
 
 // World-frame gravitational acceleration field. CentralGravityField (1/r² falloff toward planet
 // centre at {0,-R,0}) is the production implementation. Plug in a custom field via
-// FlightIntegrator::setGravityField() for exotic planets. Kept on float[3] to match the integrator's
-// GLM-free state vector.
+// FlightIntegrator::setGravityField() for exotic planets.
 struct IGravityField {
     virtual ~IGravityField() = default;
     // Gravitational acceleration (m/s^2) in the world frame [x=forward, y=up, z=right] at a world
-    // position.
-    virtual std::array<float, 3> accelWorld(const float pos_world[3]) const = 0;
-    // Geodetic (pressure) altitude (m) at a world position.
-    // Default: world-Y (correct for a flat planet with the sea-level plane at Y=0).
+    // position. Return stays float — acceleration magnitudes are small and float precision suffices.
+    virtual std::array<float, 3> accelWorld(const double pos_world[3]) const = 0;
+    // Geodetic (pressure) altitude (m) at a world position. Returns double for accuracy at
+    // planet-radius scale. Default: world-Y (flat planet, sea-level plane at Y=0).
     // CentralGravityField overrides this to return the true MSL altitude.
-    virtual float geodeticAltitude(const float pos_world[3]) const {
+    virtual double geodeticAltitude(const double pos_world[3]) const {
         return pos_world[1];
     }
 };
